chore: add cxx flags and fix issues with it

+++ ex00/megaphone.cpp
@@ -7,7 +7,7 @@ int main(const int argc, char *argv[]) {
         std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *";
     }
     for (int i = 1; i < argc; ++i) {
-        for (int j = 0; j < std::strlen(argv[i]); ++j) {
+        for (size_t j = 0; j < std::strlen(argv[i]); ++j) {
             std::putchar(std::toupper(argv[i][j]));
         }
     }
